// src/Types/addressTypes.tsx
export interface AddressType {
    firstName: string;
    lastName: string;
    companyName?: string;
    country: string;
    streetAddress1: string;
    streetAddress2?: string;
    city: string;
    state: string;
    zip: string;
    phone: string;
    email: string;
    description?: string;
}